Move function into separated file
